network module c api into shared-module
